Join configured channels when invited to them
When invited by anyone to join a channel which the bot is supposed to be
auto-joined to, the bot will now attempt to join that channel once,
automatically.

Useful for if an operator without bot access removes or otherwise excludes
the bot from the channel such that the polite auto-rejoin fails to join
the bot back to the channel, any operator from that channel can still
bring the bot back into the channel with an invitation.

Implements #8