URL_Title: clear globals before parsing
This patch prevents accidental leakage of content-type header and charset
between calls to get_title. Without clearing these, it is possible for a URL
title to be decoded from the wrong charset if a URL was previously titled
with a differing charset to the current one. This patch clears these stale
values to guarantee accurate charset decoding per URL.